refactor(go/webui): share formatKeyPreview so log key previews are masked
The request-log table showed the raw key_preview (12 leading + 6 trailing
chars, no mask), inconsistent with the API-keys page which renders it masked
via formatKeyPreview. Move formatKeyPreview into lib/format and reuse it in
both the API-keys page and the logs table, so a preview reads the same
everywhere ("sk-abcdef****…****klmnop") instead of looking like a full key.
